issue4: Remove needless comment
Since it is obvious that the slice bounds chosen are due to the
structure of the data in the CSV file, the comment was not needed.

Plus, the changes to be made to that function in the next commit
will render the comment inaccurate--removing an unnecessary comment
is easier (and better too) than updating it.