issue4: Make raw_timetable() return the raw coursenames table too
Instead of returning just the raw timetable from raw_timetable(),
return the raw coursenames table too. To make the coursenames.json
file optional--as will be done in the next commit--this behaviour is
required.

And of course, rename raw_timetable() to raw_data() since the
function now returns other data as well.