issue4: If not found, autogenerate coursenames.json
Previously, the user needed to define a coursenames.json file. Now, if
a coursenames.json is not found, it is autogenerated based on the data
in the CSV file. The user can, if s/he wants shorter coursenames to be
displayed, edit the coursenames.json.

There still is a problem though: full coursenames may not fit in
the default cell dimensions of Google Sheets. So, there needs to be
some way to adjust the cell sizes as per the text in them.