courseinfo: Use full coursenames if no name given in interactive mode
When the coursenames subcommand is used in interactive mode, its prompt
says, "If no alternative name is given, the default name is kept."

But that is not being done. Instead, when no name is given, an empty
string is used.

Fix it by using full coursename(s) when name(s) omitted.